+++ src/main/java/com/datvm/hairbookingapp/dto/request/ServicesUpdateRequest.java
@@ -27,7 +27,7 @@ public class ServicesUpdateRequest {
     String duration;
 
     @NotNull(message = "price can't be null")
-    BigDecimal price;
+    int price;
 
     @NotBlank(message = "image can't be blank. Please add image url")
     String image;
