+++ src/main/java/com/datvm/hairbookingapp/dto/response/ServicesResponse.java
@@ -16,7 +16,7 @@ public class ServicesResponse {
     String serviceName;
     String description;
     String duration;
-    BigDecimal price;
+    int price;
     String image;
     Category categories;
